# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  handlers.BulkCreateRequest:
    properties:
      products:
        items:
          $ref: '#/definitions/handlers.CreateProductRequest'
        type: array
    required:
    - products
    type: object
  handlers.CreateProductRequest:
    properties:
      description:
        example: High-performance gaming laptop
        type: string
      name:
        example: Gaming Laptop
        type: string
      price:
        example: 1500
        type: number
      stock:
        example: 10
        minimum: 0
        type: integer
    required:
    - name
    - price
    type: object
  handlers.ErrorDetail:
    properties:
      code:
        example: NOT_FOUND
        type: string
      message:
        example: Product with id 999 not found
        type: string
    type: object
  handlers.ErrorMeta:
    properties:
      timestamp:
        example: "2026-01-16T14:23:41.761Z"
        type: string
      trace_id:
        example: b79b7627-fc60-4a43-8215-efb4739f7d1d
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/handlers.ErrorDetail'
      meta:
        $ref: '#/definitions/handlers.ErrorMeta'
    type: object
  handlers.MetaResponse:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 14
        type: integer
      total_pages:
        example: 2
        type: integer
      trace_id:
        example: 1234555xxxx
        type: string
    type: object
  handlers.PaginatedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.ProductResponse'
        type: array
      meta:
        $ref: '#/definitions/handlers.MetaResponse'
    type: object
  handlers.ProductResponse:
    properties:
      created_at:
        example: "2026-01-28T13:35:27.813525Z"
        type: string
      description:
        example: High-performance gaming laptop
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Gaming Laptop
        type: string
      price:
        example: 1500
        type: number
      stock:
        example: 10
        type: integer
      updated_at:
        example: "2026-01-28T13:35:27.813525Z"
        type: string
    type: object
  handlers.SingleProductResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.ProductResponse'
      meta:
        $ref: '#/definitions/handlers.SuccessMeta'
    type: object
  handlers.SuccessMeta:
    properties:
      trace_id:
        example: 815ad63d-26ff-4940-83dd-c74a115187b9
        type: string
    type: object
  handlers.UpdateProductRequest:
    properties:
      description:
        example: High-performance gaming laptop
        type: string
      name:
        example: Gaming Laptop
        type: string
      price:
        example: 1800
        type: number
      stock:
        example: 5
        minimum: 0
        type: integer
    required:
    - name
    - price
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A Product CRUD REST API built with Go, Gin, GORM, and PostgreSQL
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go REST API
  version: "1.0"
paths:
  /products:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all products
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product with the provided information
      parameters:
      - description: Product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.SingleProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete a product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Retrieve a single product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SingleProductResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get a product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product with the provided information
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SingleProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update a product
      tags:
      - products
  /products/bulk:
    post:
      consumes:
      - application/json
      description: Create multiple products in parallel using goroutines
      parameters:
      - description: Array of products
        in: body
        name: products
        required: true
        schema:
          $ref: '#/definitions/handlers.BulkCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/handlers.ProductResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Bulk create products
      tags:
      - products
  /products/multiple:
    get:
      consumes:
      - application/json
      description: Retrieve multiple products by their IDs in parallel using goroutines
      parameters:
      - description: Comma-separated product IDs
        example: 1,2,3
        in: query
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.ProductResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get multiple products by IDs
      tags:
      - products
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"